requirements.txt
fastapi
uvicorn
celery
redis

sqlalchemy
psycopg2-binary

# ✅ NEW
reportlab

python-jose
passlib[bcrypt]
python-multipart

bcrypt==4.0.1
passlib==1.7.4